Simplify the selectors TypeScript types
