+++ packages/core-data/src/entity-types/plugin.ts
@@ -9,7 +9,6 @@ import type {
 } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -75,5 +74,6 @@ declare module './base-entity-records' {
 }
 
 export type PluginStatus = 'active' | 'inactive';
-export type Plugin< C extends Context = DefaultContextOf< 'root', 'plugin' > > =
-	OmitNevers< _BaseEntityRecords.Plugin< C > >;
+export type Plugin< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.Plugin< C >
+>;
