+++ packages/core-data/src/entity-types/post.ts
@@ -13,7 +13,6 @@ import type {
 } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -149,5 +148,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type Post< C extends Context = DefaultContextOf< 'postType', 'post' > > =
-	OmitNevers< _BaseEntityRecords.Post< C > >;
+export type Post< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.Post< C >
+>;
