+++ packages/core-data/src/entity-types/settings.ts
@@ -9,7 +9,6 @@ import type {
 } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -94,5 +93,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type Settings< C extends Context = DefaultContextOf< 'root', 'site' > > =
-	OmitNevers< _BaseEntityRecords.Settings< C > >;
+export type Settings< C extends Context = 'view' > = OmitNevers<
+	_BaseEntityRecords.Settings< C >
+>;
