+++ packages/core-data/src/entity-types/taxonomy.ts
@@ -4,7 +4,6 @@
 import type { Context, ContextualField, OmitNevers } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -88,6 +87,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type Taxonomy<
-	C extends Context = DefaultContextOf< 'postType', 'taxonomy' >
-> = OmitNevers< _BaseEntityRecords.Taxonomy< C > >;
+export type Taxonomy< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.Taxonomy< C >
+>;
