+++ packages/core-data/src/entity-types/theme.ts
@@ -4,7 +4,6 @@
 import type { Context, PostFormat, RenderedText, OmitNevers } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -222,5 +221,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type Theme< C extends Context = DefaultContextOf< 'root', 'theme' > > =
-	OmitNevers< _BaseEntityRecords.Theme< C > >;
+export type Theme< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.Theme< C >
+>;
