+++ packages/core-data/src/entity-types/user.ts
@@ -9,7 +9,6 @@ import type {
 } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -110,5 +109,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type User< C extends Context = DefaultContextOf< 'root', 'user' > > =
-	OmitNevers< _BaseEntityRecords.User< C > >;
+export type User< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.User< C >
+>;
