+++ packages/core-data/src/entity-types/widget-type.ts
@@ -4,7 +4,6 @@
 import type { Context, OmitNevers } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -33,6 +32,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type WidgetType<
-	C extends Context = DefaultContextOf< 'root', 'widgetType' >
-> = OmitNevers< _BaseEntityRecords.WidgetType< C > >;
+export type WidgetType< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.WidgetType< C >
+>;
