+++ packages/core-data/src/entity-types/widget.ts
@@ -4,7 +4,6 @@
 import type { Context, ContextualField, OmitNevers } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -60,5 +59,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type Widget< C extends Context = DefaultContextOf< 'root', 'widget' > > =
-	OmitNevers< _BaseEntityRecords.Widget< C > >;
+export type Widget< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.Widget< C >
+>;
