+++ packages/core-data/src/entity-types/wp-template.ts
@@ -10,7 +10,6 @@ import type {
 } from './helpers';
 
 import type { BaseEntityRecords as _BaseEntityRecords } from './base-entity-records';
-import type { DefaultContextOf } from './index';
 
 declare module './base-entity-records' {
 	export namespace BaseEntityRecords {
@@ -90,6 +89,6 @@ declare module './base-entity-records' {
 	}
 }
 
-export type WpTemplate<
-	C extends Context = DefaultContextOf< 'postType', 'wp_template' >
-> = OmitNevers< _BaseEntityRecords.WpTemplate< C > >;
+export type WpTemplate< C extends Context = 'edit' > = OmitNevers<
+	_BaseEntityRecords.WpTemplate< C >
+>;
