+++ packages/core-data/tsconfig.json
@@ -1,0 +1,23 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"declarationDir": "build-types",
+		"noUnusedParameters": false,
+		"checkJs": false
+	},
+	"references": [
+		{ "path": "../api-fetch" },
+		{ "path": "../data" },
+		{ "path": "../deprecated" },
+		{ "path": "../element" },
+		{ "path": "../html-entities" },
+		{ "path": "../i18n" },
+		{ "path": "../is-shallow-equal" },
+		{ "path": "../url" }
+	],
+	"include": [
+		"src/**/*.ts",
+		"src/**/*.js"
+	]
+}
